requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-multipart>=0.0.6
pydantic>=2.5.0
email-validator>=2.1.0
google-generativeai>=0.3.0
requests>=2.31.0 